Move MockHTTPClient to testsupport pkg
